+++ lib/authentication_view.dart
@@ -13,17 +13,19 @@ class AuthenticationView extends StatelessWidget{
   ScreenType screenType = ScreenType.LOGIN;
   List<FieldType> fieldTypes = [FieldType.MOBILE, FieldType.PASSWORD];
   ValidationCallback onValidation;
+  String buttonText = "CONTINUE";
+  Widget headerLayout = Container();
 
-  AuthenticationView({this.screenType, this.fieldTypes, @required this.onValidation});
+  AuthenticationView({this.screenType, this.fieldTypes, @required this.onValidation, this.buttonText, this.headerLayout});
 
   @override
   Widget build(BuildContext context) {
    switch(screenType) {
       case ScreenType.LOGIN:
-        return LoginView(fieldTypes: fieldTypes, onValidation: onValidation);
+        return LoginView(fieldTypes: fieldTypes, onValidation: onValidation, buttonText: buttonText, headerLayout: headerLayout);
 
       case ScreenType.SIGNUP:
-        return LoginView(fieldTypes: fieldTypes, onValidation: onValidation);
+        return LoginView(fieldTypes: fieldTypes, onValidation: onValidation, buttonText: buttonText, headerLayout: headerLayout);
     }
    return null;
   }