+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-## [0.0.9] - 31/07/2019 09:38 PM.
+## [0.1.0] - 31/07/2019 11:10 PM.
 
-* Provides text field values as callbacks.
+* Provides header layout option.
+* Makes button text as dynamic.